Move options-info back to app/scripts to avoid circular dependency
