+++ app/scripts/SeriesListMenu.jsx
@@ -6,7 +6,8 @@ import ContextMenuItem from './ContextMenuItem';
 import NestedContextMenu from './NestedContextMenu';
 
 // Configs
-import { OPTIONS_INFO, THEME_DARK, TRACKS_INFO_BY_TYPE } from './configs';
+import { THEME_DARK, TRACKS_INFO_BY_TYPE } from './configs';
+import OPTIONS_INFO from './options-info';
 
 // Styles
 import classes from '../styles/ContextMenu.module.scss';
